+++ src/main/groovy/com/trevorism/testing/model/TestMetadata.groovy
@@ -3,6 +3,7 @@ package com.trevorism.testing.model
 class TestMetadata {
 
     String id
+    String testSuiteId
     boolean disabled
     boolean shouldFail