+++ src/main/groovy/com/trevorism/testing/service/DefaultTestExecutorService.groovy
@@ -10,6 +10,7 @@ import com.trevorism.https.AppClientSecureHttpClient
 import com.trevorism.https.SecureHttpClient
 import com.trevorism.testing.model.TestError
 import com.trevorism.testing.model.TestEvent
+import com.trevorism.testing.model.TestMetadata
 import com.trevorism.testing.model.TestSuite
 import com.trevorism.testing.model.WorkflowRequest
 import org.slf4j.Logger
@@ -22,17 +23,25 @@ class DefaultTestExecutorService implements TestExecutorService {
     private GithubClient githubClient
     private Repository<TestSuite> testSuiteRepository
     private Repository<TestError> errorRepository
+    private TestMetadataService testMetadataService
 
     DefaultTestExecutorService() {
         SecureHttpClient appClientSecureHttpClient = new AppClientSecureHttpClient()
         githubClient = new DefaultGithubClient(appClientSecureHttpClient)
         testSuiteRepository = new FastDatastoreRepository<>(TestSuite, appClientSecureHttpClient)
         errorRepository = new FastDatastoreRepository<>(TestError, appClientSecureHttpClient)
+        testMetadataService = new DefaultTestMetadataService(appClientSecureHttpClient)
     }
 
     @Override
     boolean executeTestSuite(TestSuite testSuite) {
         String testType = testSuite.kind.toLowerCase()
+        TestMetadata metadata = testMetadataService.getMetadataByTestSuiteId(testSuite.id)
+        if(metadata && metadata.disabled){
+            log.info("Test suite ${testSuite.id} is disabled")
+            return false
+        }
+
         boolean result = githubClient.invokeWorkflow(testSuite.source, new WorkflowRequest(workflowInputs: ["TEST_TYPE": testType]))
         return result
     }
@@ -56,9 +65,13 @@ class DefaultTestExecutorService implements TestExecutorService {
         suite.lastRuntimeSeconds = (long) (testEvent.durationMillis / 1000)
 
         TestSuite updated = testSuiteRepository.update(suite.id, suite)
-
         if(!updated.lastRunSuccess)
         {
+            TestMetadata metadata = testMetadataService.getMetadataByTestSuiteId(suite.id)
+            if(metadata && (metadata.shouldFail || metadata.disabled)){
+                log.info("Test suite ${suite.id} is supposed to fail or is disabled")
+                return updated
+            }
             TestError error = new TestError(source: suite.source, message: "Test suite run failed", date: testEvent.date, details: [kind: testEvent.kind])
             errorRepository.create(error)
         }